feat: add custom hook useProducts
