+++ src/components/product/ProductList.jsx
@@ -1,47 +1,16 @@
-import { useEffect, useRef, useState, useCallback, useMemo } from 'react';
+import { useRef, useState, useCallback } from 'react';
 
 import LoadingSpinner from '../UI/LoadingSpinner';
 
 import styles from './ProductList.module.css';
 import ProductListItem from './ProductListItem';
-import config from '../../config/constants';
+import useProducts from '../../hooks/useProducts';
 
 const ProductList = () => {
-  const [productList, setProductList] = useState([]);
-  const [isLoading, setIsLoading] = useState(false);
   const [search, setSearch] = useState('');
+  const { filteredProducts, isLoading } = useProducts(search);
 
   const searchInput = useRef(null);
-
-  const fetchProductsHandler = useCallback(async () => {
-    setIsLoading(true);
-
-    try {
-      const response = await fetch(`${config.API_URL}/product`);
-      if (!response.ok) {
-        throw new Error('Something went wrong!');
-      }
-
-      const devicesData = await response.json();
-
-      setProductList(devicesData);
-    } catch (error) {
-      console.error(error.message);
-    }
-
-    setIsLoading(false);
-  }, []);
-
-  useEffect(() => {
-    fetchProductsHandler();
-  }, [fetchProductsHandler]);
-
-  const filteredProducts = useMemo(() =>
-    productList.filter((product) => {
-      const brandAndModel = `${product.brand}-${product.model}`;
-      return brandAndModel.toLowerCase().includes(search.toLowerCase());
-    }),
-  );
 
   const handleSearch = useCallback(() => {
     setSearch(searchInput.current.value);
