+++ src/config/constants.js
@@ -1,0 +1,7 @@
+const API_URL = 'https://front-test-api.herokuapp.com/api';
+
+const config = {
+  API_URL,
+};
+
+export default config;
